requirements.txt
requests>=2.31.0
docker>=7.0.0
pyyaml>=6.0
